// src/app/model/mis-report/prepaid-cust-data.model.ts
export interface PrepaidCustDataModel {
    locationCode: string;
    office: string;
    billGroup: string;
    book: string;
    consumerNo: number;
    noOfConsumer:number;
    wlkOr: string;
    pvAc: string;
    name: string;
    fatherName: string;
    addr: string;
    likely: string;
    tariff: string;
    lastBillMonth: string;
    principalArrear: number;
    lpsArrear: number;
    vatArrear: number;
    totalArrear: number;
}